+++ mslicer/src/app/project/mod.rs
@@ -1,6 +1,9 @@
 use std::path::PathBuf;
 
-use crate::app::project::model::Model;
+use crate::app::{
+    project::model::Model,
+    task::{FileDialog, ProjectLoad, ProjectSave, Task},
+};
 use common::{
     progress::CombinedProgress,
     slice::{DynSlicedFile, SliceConfig},
@@ -30,6 +33,40 @@ impl Project {
             path: Some(path),
             ..self
         }
+    }
+
+    pub fn reset(&mut self, default_config: &SliceConfig) {
+        self.path = None;
+        self.slice_config = default_config.clone();
+        self.post_processing = Default::default();
+        self.models.clear();
+    }
+}
+
+impl Project {
+    pub fn load() -> FileDialog {
+        FileDialog::pick_file(("mslicer project", &["mslicer"]), |_app, path, tasks| {
+            tasks.push(Box::new(ProjectLoad::new(path.to_path_buf())))
+        })
+    }
+
+    pub fn save(&self) -> Box<dyn Task> {
+        if let Some(path) = self.path.clone() {
+            Box::new(ProjectSave::new(self.clone(), path.to_path_buf()))
+        } else {
+            Box::new(self.save_as())
+        }
+    }
+
+    pub fn save_as(&self) -> FileDialog {
+        FileDialog::save_file(("mslicer project", &["mslicer"]), |app, path, tasks| {
+            let path = path.with_extension("mslicer");
+            tasks.push(Box::new(ProjectSave::new(
+                app.project.clone(),
+                path.to_path_buf(),
+            )));
+            app.project.path = Some(path);
+        })
     }
 }