+++ mslicer/src/ui/popup.rs
@@ -8,8 +8,8 @@ use egui_phosphor::regular::X;
 
 use crate::{
     app::{
-        App, is_slicing, project::Project, remote_print::RemotePrint,
-        slice_operation::SliceOperation,
+        App, config::Config, is_slicing, project::Project, remote_print::RemotePrint,
+        slice_operation::SliceOperation, task::TaskManager,
     },
     app_ref_type,
     ui::{panels::Panels, state::UiState},
@@ -35,9 +35,11 @@ pub struct Popup {
 
 pub struct PopupApp<'a> {
     pub panels: &'a mut Panels,
+    pub tasks: &'a mut TaskManager,
     pub remote_print: &'a mut RemotePrint,
     pub slice_operation: &'a mut Option<SliceOperation>,
     pub state: &'a mut UiState,
+    pub config: &'a mut Config,
     pub project: &'a mut Project,
 }
 
@@ -63,9 +65,11 @@ impl<'a> PopupManagerRef<'a> {
         let this = &mut self.app.popup;
         let mut app = PopupApp {
             panels: &mut self.app.panels,
+            tasks: &mut self.app.tasks,
             remote_print: &mut self.app.remote_print,
             slice_operation: &mut self.app.slice_operation,
             state: &mut self.app.state,
+            config: &mut self.app.config,
             project: &mut self.app.project,
         };