+++ Changelog.md
@@ -26,6 +26,7 @@
 - Pick between Perspective or Orthographic camera projections
 - Camera pitch is no longer locked between (-90°, 90°)
 - Exposure test generator ([#24](https://github.com/connorslade/mslicer/pull/24))
+- Button to create new project
 
 ## v0.5.0 &mdash; February 21st, 2026